Update From.cs
Apologies, I accidentally overwrote the Phone property in my previous change. When the Ticket was created as a follow-up to a closed ticket, the ticket_id property is specified in the From property coming back from the API. Is this all that is needed to fill out the TicketId property here?